refactor importFidesPackageVersion to be shared
